pyproject.toml
[project]
name = "elastic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=9.3.0",
    "flask>=3.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
